pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.9.2",
    "openai-agents>=0.6.1",
    "python-dotenv>=1.2.1",
]
